pyproject.toml
[tool.poetry]
name = "powerswitch-pi"
version = "0.1.0"
description = ""
authors = ["Max Gonzih <gonzih@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
gpiozero = "^1.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
